// engineering/packages/domains/siakad/src/submodules/awards/src/domain/types/AwardsEnums.ts
export enum AwardCategory {
  BEST_RESEARCH = 'BEST_RESEARCH',
  BEST_PUBLICATION = 'BEST_PUBLICATION',
  BEST_REVIEWER = 'BEST_REVIEWER',
  BEST_PRESENTATION = 'BEST_PRESENTATION',
  BEST_INNOVATION = 'BEST_INNOVATION',
  LIFETIME_ACHIEVEMENT = 'LIFETIME_ACHIEVEMENT',
  YOUNG_ACADEMICIAN = 'YOUNG_ACADEMICIAN',
  OUTSTANDING_CONTRIBUTION = 'OUTSTANDING_CONTRIBUTION'
}

export enum AwardCycle {
  ANNUAL = 'ANNUAL',
  BIENNIAL = 'BIENNIAL',
  SPECIAL = 'SPECIAL'
}

export enum AwardStatus {
  DRAFT = 'DRAFT',
  NOMINATION_OPEN = 'NOMINATION_OPEN',
  EVALUATION_PHASE = 'EVALUATION_PHASE',
  DECIDED = 'DECIDED',
  PUBLISHED = 'PUBLISHED',
  CANCELED = 'CANCELED'
}

export enum SourceContext {
  PUBLICATION = 'PUBLICATION',
  RESEARCH = 'RESEARCH',
  CONFERENCE = 'CONFERENCE',
  CERTIFICATION = 'CERTIFICATION',
  OTHER = 'OTHER'
}

export enum RecommendationType {
  STRONGLY_RECOMMEND = 'STRONGLY_RECOMMEND',
  RECOMMEND = 'RECOMMEND',
  NEUTRAL = 'NEUTRAL',
  NOT_RECOMMEND = 'NOT_RECOMMEND'
}

export enum AwardDecisionType {
  WINNER = 'WINNER',
  RUNNER_UP = 'RUNNER_UP',
  HONORABLE_MENTION = 'HONORABLE_MENTION',
  NOT_SELECTED = 'NOT_SELECTED'
}